extract OptionsMethodRequest class
